more updates for install scripts

+++ scripts/install_p2.sh
@@ -14,11 +14,19 @@ echo "**** Starting jetty"
 sudo service jetty8 start
 
 echo "**** Replacing SOLR default schema with CKAN schema"
-sudo mv /etc/solr/conf/schema.xml /etc/solr/conf/schema.xml.bak
-sudo ln -s /usr/lib/ckan/default/src/ckan/ckan/config/solr/schema.xml
+if [ ! -f /etc/solr/conf/schema.xml.bak ]; then
+  echo "**** Backing up old schema file"
+  sudo mv /etc/solr/conf/schema.xml /etc/solr/conf/schema.xml.bak
+fi
+
+if [ -f /etc/solr/conf/schema.xml ]; then
+  rm /etc/solr/conf/schema.xml
+fi
+
+sudo ln -s /usr/lib/ckan/default/src/ckan/ckan/config/solr/schema.xml /etc/solr/conf/schema.xml
 
 echo "**** Restarting jetty"
-sudo service jetty restart
+sudo service jetty8 restart
 
 echo "**** Creating PG tables"
 . /usr/lib/ckan/default/bin/activate
@@ -26,7 +34,9 @@ cd /usr/lib/ckan/default/src/ckan
 paster db init -c /etc/ckan/default/development.ini
 
 echo "**** Linking who.ini"
-ln -s /usr/lib/ckan/default/src/ckan/who.ini /etc/ckan/default/who.ini
+if[ ! -f /etc/ckan/default/who.ini ]; then
+  ln -s /usr/lib/ckan/default/src/ckan/who.ini /etc/ckan/default/who.ini
+fi
 
 echo "**** Starting CKAN"
 echo ". /usr/lib/ckan/default/bin/activate"
